pyproject.toml
[project]
name = "ag-news"
version = "0.1.0"
description = "Classification task"
authors = [
    {name = "shamil.a",email = "shamil.a@phystech.edu"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "pytorch-lightning (>=2.5.1,<3.0.0)",
    "optuna (>=4.2.1,<5.0.0)",
    "optuna-integration (>=4.2.1,<5.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "lightning (>=2.5.1,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "torch (==2.7.0)",
    "torchtext (==0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
fire = "^0.7.0"

